2.0.27.4
Changes:

* add PDDeviceN - Special Color Space
* update annotation processor
* update PlainText layout
* update base models to 3.x version
* fix PDIndexed Color Space
* fix from1Bit bitmap conveter
* fix an issue with the path stroke width
* fix minor issues in SampledImageReader

+++ library/src/test/java/com/tom_roush/pdfbox/cos/TestCOSInteger.java
@@ -104,7 +104,7 @@ public class TestCOSInteger extends TestCOSNumber
     {
         for (int i = -1000; i < 3000; i += 200)
         {
-            assertEquals((double) i, COSInteger.get(i).doubleValue());
+            assertEquals((float) i, COSInteger.get(i).floatValue());
         }
     }
